Bunch of fixes and updates. Inhibitted trees are no longer deleted after
Bunch of fixes and updates.  Inhibitted trees are no longer deleted after
removal, but are saved until the namespace is freed.  This should allow
full traversal of the trees from the namespace for extra info, without
having them in the primary tree.

If you have any problems you can use the V0_2_1 tag in the meantime.